[pf-579] Rename catAlgorithm.js to dash-case
